User restructure
Made a User service
