+++ VitoshaBank/VitoshaBank/Program.cs
@@ -1,15 +1,9 @@
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace VitoshaBank
 {
-	public class Program
+    public class Program
 	{
 		public static void Main(string[] args)
 		{
