--- VitoshaBank/VitoshaBank/Services/Interfaces/IUserService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace VitoshaBank.Services.Interfaces
-{
-    public interface IUserService
-    {
-    }
-}
